testando exclusao

+++ painel/paginas/usuarios/excluir.php
@@ -1,4 +1,5 @@
 <?php 
+session_start();
 $tabela = 'usuarios';
 require_once("../../../conexao.php");
 
@@ -12,8 +13,16 @@ if($foto != "sem-foto.jpg"){
 	@unlink('../../images/perfil/'.$foto);
 }
 
+if($nivel != "Administrador"){
+	
 // Primeiro, excluir as ações relacionadas ao usuário
 $pdo->query("DELETE FROM acao_realizada WHERE fk_usuarios_id = $id");
 
+
+
 $pdo->query("DELETE FROM $tabela WHERE id = '$id' ");
 echo 'Excluído com Sucesso';
+
+	}else{
+		alert("Usuario Administrador não pode ser excluido!");
+}
